refactor: centralize credit price logic into custom hook
