+++ frontend/src/components/SummaryCard.tsx
@@ -1,52 +1,15 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useSearchParams } from "next/navigation";
 import Image from 'next/image'
 import { Card, CardContent } from '@/components/ui/card'
-import { getCreditPrice, type CreditPrice } from '@/services/creditService'
+import { useCreditPrice } from '@/hooks/useCreditPrice'
 
 interface SummaryCardProps {
   isDesktop?: boolean
 }
 
 export function SummaryCard({ isDesktop = false }: Readonly<SummaryCardProps>) {
-  const searchParams = useSearchParams();
-  const [creditPrice, setCreditPrice] = useState<CreditPrice | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  // Extract parameters from URL
-  const co2Quantity = Number(searchParams.get('co2')) || 1;
-  const creditPriceId = searchParams.get('cred') || '2';
-
-  useEffect(() => {
-    async function fetchCreditPrice() {
-      try {
-        setLoading(true);
-        setError(null);
-        const price = await getCreditPrice(creditPriceId);
-        setCreditPrice(price);
-      } catch (err) {
-        setError('Erro ao carregar preço');
-        console.error('Error fetching credit price:', err);
-      } finally {
-        setLoading(false);
-      }
-    }
-
-    fetchCreditPrice();
-  }, [creditPriceId]);
-
-  const unitPrice = creditPrice?.amout ?? 0;
-  const totalValue = co2Quantity * unitPrice;
-
-  const formatCurrency = (value: number) => {
-    return new Intl.NumberFormat('pt-BR', {
-      style: 'currency',
-      currency: 'BRL'
-    }).format(value);
-  };
+  const { loading, error, totalValue, formatCurrency } = useCreditPrice();
 
   const displayValue = loading ? 'Carregando...' : error ? 'Erro' : formatCurrency(totalValue);
